+++ custom_typings/path-is-inside.d.ts
@@ -1,0 +1,4 @@
+declare module 'path-is-inside' {
+  function pathIsInside(thePath: string, potentialParent: string): boolean;
+  export = pathIsInside;
+}