+++ src/url-loader/fs-url-loader.ts
@@ -14,9 +14,9 @@
 
 import * as fs from 'fs';
 import * as pathlib from 'path';
+import pathIsInside = require('path-is-inside');
 import Uri from 'vscode-uri';
 
-import {isPathInside} from '../core/utils';
 import {ResolvedUrl} from '../index';
 import {Result} from '../model/analysis';
 import {PackageRelativeUrl} from '../model/url';
@@ -30,15 +30,13 @@ export class FSUrlLoader implements UrlLoader {
   root: string;
 
   constructor(root: string = '') {
-    if (root.endsWith('/')) {
-      root += '/';
-    }
     this.root = pathlib.resolve(root);
   }
 
   canLoad(url: ResolvedUrl): boolean {
-    return url.startsWith('file://') &&
-        isPathInside(this.root, Uri.parse(url).fsPath);
+    const parsed = Uri.parse(url);
+    return parsed.scheme === 'file' && !parsed.authority &&
+        pathIsInside(parsed.fsPath, this.root);
   }
 
   load(url: ResolvedUrl): Promise<string> {